+++ Source/Chomp/Private/Pawns/Movement/MoveInDirectionResult.cpp
@@ -1,0 +1,1 @@
+#include "MoveInDirectionResult.h"
