+++ Source/Chomp/Private/Pawns/Movement/MovementResult.h
@@ -1,0 +1,7 @@
+#pragma once
+
+struct FMovementResult
+{
+	FVector NewLocation;
+	FRotator NewRotation;
+};
