+++ Source/Chomp/Private/Pawns/Movement/PeriodicDotProductResult.cpp
@@ -1,0 +1,1 @@
+#include "PeriodicDotProductResult.h"
