+++ Source/Chomp/Private/Pawns/Movement/PeriodicDotProductResult.h
@@ -1,0 +1,7 @@
+#pragma once
+
+struct FPeriodicDotProductResult
+{
+	bool MovedPastTarget{false};
+	double AmountMovedPast{0.0};
+};
